module1_ollama_basics/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "module1-ollama-basics"
version = "0.1.0"
description = "Module 1: Ollama Basics - Direct API interaction with Ollama using httpx"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
]
